Applying CSS to the project
Started applying basic CSS to the project. I'm going for the theme in
Valheim so that it looks the part.

I'm also planning to make it into an ingredients planner.
Just put down the dishes you want to make, the quantity,
and it'll compute the different ingredients you'll need plus how many
of each ingredient you'll be needing. At least to add some funtionality
and not just have it as a database of recipes.